Refactored the test suits for the  object
